atari

+++ utils/pos.c
@@ -6,14 +6,14 @@
 /*   By: ncolomer <ncolomer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/31 15:55:48 by ncolomer          #+#    #+#             */
-/*   Updated: 2019/10/31 22:25:06 by ncolomer         ###   ########.fr       */
+/*   Updated: 2019/11/01 12:00:08 by ncolomer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils.h"
 
 void
-	set_pos(t_pos *pos, int x, int y)
+	set_pos(t_pos *pos, double x, double y)
 {
 	pos->x = x;
 	pos->y = y;
